refactors backend tests for datatree IO
